Log but tolerate liveries that do not parse.
Failing in this step means campaign generation will fail as well, thus we should always handle this gracefully. In this particular we're dealing with a new livery that sticks out in the sense that it has a "custom_args" section.
Another scenario can occur where one of the liveries gets corrupted, for whatever reason, in which case we'd also want to skip the livery instead.

Co-Authored-By: Dan Albert <danalbert@google.com>