Fix AttributeError in Point.__eq__.
The other type this is compared to might not be Point, and that
shouldn't raise an error. It's often a programming mistake to write that
comparison, but there's a way to run into this without the caller having
a bug: comparing two instances of a dataclass with an `Point | None`
member where the rhs has None and the lhs has a Point.