Data export for DCS 2.9.2.49629 Open Beta.
This includes new AIM-9s and rockets for the F-16, and JSOW for the
F-15E.

The M-2000 no longer has some of the properties for configuring LGB
codes and gun/rocket ripple settings which looks odd, but the mission
editor agrees. Those are configurable via the kneeboard in cockpit so I
suppose razbam doesn't think they're needed any more.

There are some other changes as well but those are the highlights.